+++ apps/everest/components/time-selection/time-selection.utils.ts
@@ -104,7 +104,7 @@ export const getFormValuesFromCronExpression = (cron: string) => {
     dayOfWeek: +parts[4],
   };
 
-  //TODO during working on EVEREST-485 move DbWizardFormField and ScheduledFormField to "TimeFields" to avoid duplicating
+  // TODO during working on EVEREST-485 move DbWizardFormField and ScheduledFormField to "TimeFields" to avoid duplicating
 
   return {
     [DbWizardFormFields.minute]: !Number.isNaN(cronObj.minute)
